requirements.txt
numpy
joblib
streamlit
scikit-learn
pillow
